research-assistant/requirements.txt
duckduckgo-search
langchain
langchain_community
langchain-openai
openai
python-dotenv
requests
beautifulsoup4
markdown
weasyprint